refactor: 장소 추출시 중복되는 장소 있을 시 중복체크 후 한 장소만 저장
